feat: show abstract service, add autofocus to inputs
## Description

I wanted to add in an abstract TodosService contract to show that we'd be able to easily refactor this to use some other datastore rather than jotai/local state.

Additionally, I added autofocus to the inputs so that when you want to create a child issue, it auto focuses the input.